Moving CommandBarDialog components to separate files
