+++ src/components/command-bar/components/dialog/body.tsx
@@ -1,0 +1,7 @@
+import s from './command-bar-dialog.module.css'
+
+const CommandBarDialogBody = () => {
+	return <div className={s.body}>body</div>
+}
+
+export default CommandBarDialogBody
