+++ src/components/command-bar/components/dialog/types.ts
@@ -1,7 +1,6 @@
 import { ReactNode } from 'react'
 
 interface CommandBarDialogProps {
-	children: ReactNode[]
 	isOpen?: boolean
 	onDismiss?: () => void
 }
